Move DAILY_NOTE_TYPE and target helpers into daily-notes plugin
Removes the last daily-notes-shaped seams from core:

- DAILY_NOTE_TYPE const + its TypeContribution move from
  src/data/blockTypes.ts → src/plugins/daily-notes/schema.ts. The
  plugin contributes the type via a new dailyNotesDataExtension wired
  into staticDataExtensions (alongside the kernel + other data-only
  plugins). KERNEL_TYPE_CONTRIBUTIONS no longer ships a daily-note
  entry — that registration is the plugin's responsibility now.

- isDateAlias, ensureDailyNoteTarget, and the daily-note namespace
  UUID move from src/data/targets.ts → daily-notes plugin. The
  duplicate DAILY_NOTE_NS const in targets.ts is gone; the plugin's
  copy is the source of truth. computeDailyNoteId (which had
  reversed arg order vs. dailyNoteBlockId) is dropped — backlinks
  uses dailyNoteBlockId(workspaceId, iso) directly now.

- Cross-plugin imports: srs-rescheduling/schema.ts imports
  DAILY_NOTE_TYPE from @/plugins/daily-notes (was @/data/blockTypes).
  backlinks/referencesProcessor.ts imports ensureDailyNoteTarget,
  isDateAlias, and dailyNoteBlockId from @/plugins/daily-notes (was
  @/data/targets).

- Tests that construct a bare Repo now have to install
  dailyNotesDataExtension explicitly via setFacetRuntime — same
  pattern srs-rescheduling tests already use for their own data
  extension. Affects daily-notes/test/dailyNotes.test.ts,
  daily-notes/test/targetIntegration.test.ts (new — migrated from
  data/targets.test.ts), backlinks/test/referencesProcessor.test.ts,
  srs-rescheduling/test/actions.test.ts,
  utils/roamImport/test/import.test.ts. The daily-note slice of
  data/targets.test.ts (computeDailyNoteId, isDateAlias,
  ensureDailyNoteTarget) is now in the plugin tree.

- Stale comments that pointed at a server-side supabase
  deterministic-seed migration are removed (the migration no longer
  depends on these client-side constants).
